+++ webextension/data/js/classes/responseDocument.js
@@ -1,0 +1,19 @@
+/**
+ * Parse response body as HTML or XML document (proper Content-Type header must be available)
+ * @return {Promise<Document|XMLDocument>}
+ */
+Response.prototype.document = async function document() {
+	const text = await this.text(),
+		domParser = new DOMParser()
+	;
+
+	let mimeType;
+	if (this.headers.has('Content-Type')) {
+		mimeType = this.headers.get('Content-Type')
+			.split(';')
+			.find(s => /^[-\w.]+\/[-\w.]+$/i.test(s.trim()))
+		;
+	}
+
+	return domParser.parseFromString(text, mimeType);
+};